refactor: return error from router, no panic
